+++ web/stockdata/services/sources/YahooFinanceClient.py
@@ -1,4 +1,4 @@
-class YahooFinanceClient():
+class YahooFinanceClient:
 
     def get_stock_info(self, symbol):
         pass
